move the snapshot code to the new file
