+++ pennylane/debugging/debugger.py
@@ -18,232 +18,9 @@ interactive debugging of quantum circuits.
 import copy
 import pdb
 import sys
-import warnings
 from contextlib import contextmanager
-from functools import partial
-from typing import Callable, Sequence
 
 import pennylane as qml
-from pennylane.tape import QuantumTape
-from pennylane.transforms import transform
-from pennylane.typing import Result, ResultBatch
-
-
-def _is_snapshot_compatible(dev):
-    # The `_debugger` attribute is a good enough proxy for snapshot compatibility
-    return hasattr(dev, "_debugger")
-
-
-class _Debugger:
-    """A debugging context manager.
-
-    Without an active debugging context, devices will not save their internal state when
-    encoutering Snapshot operations. The debugger also serves as storage for the device states.
-
-    Args:
-        dev (Device): device to attach the debugger to
-    """
-
-    def __init__(self, dev):
-        self.snapshots = {}
-        self.active = False
-        self.device = dev
-        dev._debugger = self
-
-    def __enter__(self):
-        self.active = True
-        return self
-
-    def __exit__(self, exc_type, exc_value, exc_traceback):
-        self.active = False
-        self.device._debugger = None
-
-
-@transform
-def snapshots(tape_: QuantumTape) -> tuple[Sequence[QuantumTape], Callable[[ResultBatch], Result]]:
-    r"""This transform processes :class:`~pennylane.Snapshot` instances contained in a circuit,
-    depending on the compatibility of the execution device.
-    For supported devices, the snapshots' measurements are computed as the execution progresses.
-    Otherwise, the :func:`QuantumTape <pennylane.tape.QuantumTape>` gets split into several, one for each snapshot, with each aggregating
-    all the operations up to that specific snapshot.
-
-    Args:
-        tape_ (QNode or QuantumTape or Callable): a quantum circuit.
-
-    Returns:
-        dictionary (dict) or qnode (QNode) or quantum function (Callable) or tuple[List[QuantumTape], function]:
-        The transformed circuit as described in :func:`qml.transform <pennylane.transform>`.
-
-    If tape splitting is carried out, the transform will be conservative about the wires that it includes in each tape.
-    So, if all operations preceding a snapshot in a 3-qubit circuit has been applied to only one wire,
-    the tape would only be looking at this wire. This can be overriden by the configuration of the execution device
-    and its nature.
-
-    Regardless of the transform's behaviour, the output is a dictionary where each key is either
-    the tag supplied to the snapshot or its index in order of appearance, in addition to an
-    ``"execution_results"`` entry that returns the final output of the quantum circuit. The post-processing
-    function is responsible for aggregating the results into this dictionary.
-
-    When the transform is applied to a QNode, the ``shots`` configuration is inherited from the device.
-    Therefore, the snapshot measurements must be supported by the device's nature. An exception of this are
-    the :func:`qml.state <pennylane.state>` measurements on finite-shot simulators.
-
-    .. warning::
-
-        For devices that do not support snapshots (e.g QPUs, external plug-in simulators), be mindful of
-        additional costs that you might incur due to the 1 separate execution/snapshot behaviour.
-
-    **Example**
-
-    .. code-block:: python3
-
-        dev = qml.device("default.qubit", wires=2)
-
-        @qml.qnode(dev, interface=None)
-        def circuit():
-            qml.Snapshot(measurement=qml.expval(qml.Z(0)))
-            qml.Hadamard(wires=0)
-            qml.CNOT(wires=[0, 1])
-            qml.Snapshot()
-            return qml.expval(qml.X(0))
-
-    >>> qml.snapshots(circuit)()
-    {0: 1.0,
-    1: array([0.70710678+0.j, 0.        +0.j, 0.        +0.j, 0.70710678+0.j]),
-    'execution_results': 0.0}
-
-    .. code-block:: python3
-
-        dev = qml.device("default.qubit", shots=200, wires=2)
-
-        @qml.snapshots
-        @qml.qnode(dev, interface=None)
-        def circuit():
-            qml.Hadamard(wires=0)
-            qml.Snapshot()
-            qml.CNOT(wires=[0, 1])
-            return qml.counts()
-
-    >>> circuit()
-    {0: array([0.70710678+0.j, 0.        +0.j, 0.70710678+0.j, 0.        +0.j]),
-    'execution_results': {'00': 101, '11': 99}}
-
-    Here one can see how a device that does not natively support snapshots executes two different circuits:
-
-    .. code-block:: python3
-
-        @qml.snapshots
-        @qml.qnode(qml.device("lightning.qubit", shots=100, wires=2), diff_method="parameter-shift")
-        def circuit():
-            qml.Hadamard(wires=0),
-            qml.Snapshot(qml.counts())
-            qml.CNOT(wires=[0, 1])
-            return qml.expval(qml.PauliZ(0))
-
-        with circuit.device.tracker:
-            out = circuit()
-
-    >>> circuit.device.tracker.totals
-    {'batches': 1, 'simulations': 2, 'executions': 2, 'results': 0.0}
-
-    >>> out
-    {0: {'00': tensor(51, requires_grad=True), '10': tensor(49, requires_grad=True)}, 'execution_results': tensor(0., requires_grad=True)}
-
-    Here you can see the default behaviour of the transform for unsupported devices and you can see how the amount of wires included
-    in each resulting tape is minimal:
-
-    .. code-block:: python3
-
-        ops = [
-            qml.Snapshot(),
-            qml.Hadamard(wires=0),
-            qml.Snapshot("very_important_state"),
-            qml.CNOT(wires=[0, 1]),
-            qml.Snapshot(),
-        ]
-
-        measurements = [qml.expval(qml.PauliX(0))]
-
-        tape = qml.tape.QuantumTape(ops, measurements)
-
-        tapes, collect_results_into_dict = qml.snapshots(tape)
-
-    >>> print(tapes)
-    [<QuantumTape: wires=[], params=0>, <QuantumTape: wires=[0], params=0>, <QuantumTape: wires=[0, 1], params=0>, <QuantumTape: wires=[0, 1], params=0>]
-    """
-
-    qml.devices.preprocess.validate_measurements(tape_)
-
-    new_tapes = []
-    accumulated_ops = []
-    snapshot_tags = []
-
-    for op in tape_.operations:
-        if isinstance(op, qml.Snapshot):
-            snapshot_tags.append(op.tag or len(new_tapes))
-            meas_op = op.hyperparameters["measurement"]
-
-            new_tapes.append(
-                type(tape_)(ops=accumulated_ops, measurements=[meas_op], shots=tape_.shots)
-            )
-        else:
-            accumulated_ops.append(op)
-
-    # Create an additional final tape if a return measurement exists
-    if tape_.measurements:
-        snapshot_tags.append("execution_results")
-        new_tapes.append(type(tape_)(ops=accumulated_ops, measurements=tape_.measurements))
-
-    def postprocessing_fn(results, snapshot_tags):
-        return dict(zip(snapshot_tags, results))
-
-    if len(new_tapes) > 1:
-        warnings.warn(
-            "Snapshots are not supported for the given device. Therefore, a tape will be "
-            f"created for each snapshot, resulting in a total of {len(new_tapes)} executions.",
-            UserWarning,
-        )
-
-    return new_tapes, partial(postprocessing_fn, snapshot_tags=snapshot_tags)
-
-
-@snapshots.custom_qnode_transform
-def snapshots_qnode(self, qnode, targs, tkwargs):
-    """A custom QNode wrapper for the snapshot transform :func:`~.snapshots`.
-    Depending on whether the QNode's device supports snapshots, an efficient execution
-    would be used. Otherwise, the QNode's tape would be split into several around the
-    present snapshots and execute each individually.
-    """
-
-    def get_snapshots(*args, **kwargs):
-        # Need to construct to generate the tape and be able to validate
-        qnode.construct(args, kwargs)
-        qml.devices.preprocess.validate_measurements(qnode.tape)
-
-        old_interface = qnode.interface
-        if old_interface == "auto":
-            qnode.interface = qml.math.get_interface(*args, *list(kwargs.values()))
-
-        with _Debugger(qnode.device) as dbg:
-            # pylint: disable=protected-access
-            if qnode._original_device:
-                qnode._original_device._debugger = qnode.device._debugger
-
-            results = qnode(*args, **kwargs)
-
-            # Reset interface
-            if old_interface == "auto":
-                qnode.interface = "auto"
-            if qnode._original_device:
-                qnode.device._debugger = None
-
-        dbg.snapshots["execution_results"] = results
-        return dbg.snapshots
-
-    if _is_snapshot_compatible(qnode.device):
-        return get_snapshots
-
-    return self.default_qnode_transform(qnode, targs, tkwargs)
 
 
 class PLDB(pdb.Pdb):
